+++ openrouteservice/src/main/java/heigit/ors/routing/graphhopper/extensions/flagencoders/FootFlagEncoder.java
@@ -36,7 +36,7 @@ import static com.graphhopper.routing.util.PriorityCode.*;
  */
 public abstract class FootFlagEncoder extends ORSAbstractFlagEncoder {
     static final int SLOW_SPEED = 2;
-    static final int MEAN_SPEED = 5;
+    private static final int MEAN_SPEED = 5;
     static final int FERRY_SPEED = 15;
 
     private final Set<String> safeHighwayTags = new HashSet<>();
@@ -134,6 +134,10 @@ public abstract class FootFlagEncoder extends ORSAbstractFlagEncoder {
         maxPossibleSpeed = FERRY_SPEED;
 
         init();
+    }
+
+    public double getMeanSpeed() {
+        return MEAN_SPEED;
     }
 
     @Override